fix: remove depedency cycle
