+++ packages/apps/human-app/frontend/src/api/api-client.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-cycle -- cause by refresh token retry
 import { createFetcher } from '@/api/fetcher';
 import { env } from '@/shared/env';
