+++ packages/apps/human-app/frontend/src/api/fetch-refresh-token.ts
@@ -1,6 +1,6 @@
-import { signInSuccessResponseSchema } from '@/api/services/worker/sign-in';
 import { apiPaths } from '@/api/api-paths';
 import { browserAuthProvider } from '@/shared/helpers/browser-auth-provider';
+import { signInSuccessResponseSchema } from '@/api/services/worker/sign-in/schema';
 
 export const fetchTokenRefresh = async (baseUrl: string) => {
   const response = await fetch(
