+++ packages/apps/human-app/frontend/src/api/services/worker/sign-in/sign-in.ts
@@ -1,32 +1,11 @@
-import { z } from 'zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import { t } from 'i18next';
 import { apiClient } from '@/api/api-client';
 import { apiPaths } from '@/api/api-paths';
 import { routerPaths } from '@/router/router-paths';
 import { useAuth } from '@/auth/use-auth';
-
-export const signInDtoSchema = z.object({
-  email: z.string().email(t('validation.invalidEmail')),
-  password: z
-    .string()
-    .min(1, t('validation.passwordMissing'))
-    .max(50, t('validation.max', { count: 50 })),
-  // eslint-disable-next-line camelcase -- export vite config
-  h_captcha_token: z.string().min(1, t('validation.captcha')).default('token'),
-});
-
-export type SignInDto = z.infer<typeof signInDtoSchema>;
-
-export const signInSuccessResponseSchema = z.object({
-  // eslint-disable-next-line camelcase -- data from api
-  access_token: z.string(),
-  // eslint-disable-next-line camelcase -- data from api
-  refresh_token: z.string(),
-});
-
-export type SignInSuccessResponse = z.infer<typeof signInSuccessResponseSchema>;
+import { type SignInDto } from './types';
+import { signInSuccessResponseSchema } from './schema';
 
 function signInMutationFn(data: SignInDto) {
   return apiClient.fetcher(apiPaths.worker.signIn.path, {
