+++ packages/apps/human-app/frontend/src/api/services/worker/sign-in/types.ts
@@ -1,0 +1,9 @@
+import { type z } from 'zod';
+import {
+  type signInDtoSchema,
+  type signInSuccessResponseSchema,
+} from './schema';
+
+export type SignInDto = z.infer<typeof signInDtoSchema>;
+
+export type SignInSuccessResponse = z.infer<typeof signInSuccessResponseSchema>;
